Moved class `Hex` to new file `database/map.py`
